test:전략 패턴을 위해 나눗셈 테스트를 StringCalculatorTest에서 DivideOperatorTest로 리팩토링
